almost refactored
