+++ src/components/BurgerMenu/index.ts
@@ -1,1 +1,1 @@
-export * from '../BurgerMenu';
+export * from './BurgerMenu';
