+++ src/components/Header/styles.module.scss
@@ -93,7 +93,7 @@
   height: 100%;
   box-shadow: -1px 0 0 0 $elements;
   display: block;
-  background-image: url(/src/assets/icons/burger-icon.svg);
+  background-image: url(/src/assets/images/icons/burger-icon.svg);
   background-repeat: no-repeat;
   background-position: center;
   cursor: pointer;
@@ -109,7 +109,7 @@
   height: 100%;
   box-shadow: -1px 0 0 0 $elements;
   display: block;
-  background-image: url(/src/assets/icons/close-burger-icon.svg);
+  background-image: url(/src/assets/images/icons/close-burger-icon.svg);
   background-repeat: no-repeat;
   background-position: center;
   cursor: pointer;
