+++ src/enums/index.ts
@@ -1,2 +1,3 @@
 export * from './Paths';
 export * from './PrimaryButtons';
+export * from './SortType';
