+++ src/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './usePhone';
 export * from './useAddToCart';
 export * from './useRemoveFromCart';
 export * from './useUpdateCartQuantity';
+export * from './useProducts';
