--- src/types/Tablet.ts
@@ -1,24 +1,0 @@
-export interface Tablet {
-  id: string;
-  category: string;
-  namespaceId: string;
-  name: string;
-  capacityAvailable: string[];
-  capacity: string;
-  priceRegular: number;
-  priceDiscount?: number;
-  colorsAvailable: string[];
-  color: string;
-  images: string[];
-  description: Array<{
-    title: string;
-    text: string[];
-  }>;
-  screen: string;
-  resolution: string;
-  processor: string;
-  ram: string;
-  camera: string;
-  zoom: string;
-  cell: string[];
-}
