+++ src/utils/index.ts
@@ -1,3 +1,6 @@
 export * from './getCartProducts';
 export * from './getCartProductsQuantity';
 export * from './getTotalCost';
+export * from './getSearchWith';
+export * from './scrollToTop';
+export * from './sortDevices';
